updated readme and setup
