+++ python/samples/demos/graphrag/setup-part-2.sh
@@ -1,0 +1,3 @@
+# Make sure you have added your OpenAI key to the .env file
+# and made any other changes you want to the config file
+graphrag index --root ./ragtest
